Backend: altered id from long to interger. Android: Created Register and connected to back end . IT LIVES

+++ backend/src/main/java/com/backend/lane/controllers/EParticipantsController.java
@@ -14,12 +14,12 @@ public class EParticipantsController {
 
     @PostMapping("/add")
     public String addParticipant() {
-        eParticipantsService.addParticipantToEvent(1L, 1L);
+        eParticipantsService.addParticipantToEvent(1, 1);
         return "Participant added";
     }
 
     @DeleteMapping("/{id}")
-    public String deleteParticipant(@PathVariable Long id) {
+    public String deleteParticipant(@PathVariable Integer id) {
         return "Participant deleted";
     }
 }
