+++ backend/src/main/java/com/backend/lane/service/EParticipantsService.java
@@ -2,7 +2,7 @@ package com.backend.lane.service;
 
 public interface EParticipantsService {
 
-    void addParticipantToEvent(Long event_id, Long participant_id);
-    void deleteParticipantFromEvent(Long event_id, Long participant_id);
+    void addParticipantToEvent(Integer event_id, Integer participant_id);
+    void deleteParticipantFromEvent(Integer event_id, Integer participant_id);
 
 }
